requirements.txt
Flask
python-dotenv
Pillow
moviepy
requests